requirements.txt
flask
gpt4all
google-generativeai
langchain
langchain-community
faiss-cpu 